// superset-frontend/plugins/plugin-chart-custom-filter/src/CustomControls/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import {
  DataRecord,
  FilterState,
  QueryFormColumn,
  QueryFormData,
  SetDataMaskHook,
} from '@superset-ui/core';

export type ControlType = 'Dropdown' | 'Radio' | 'Checkbox' | 'TextBox';

export type Orientation = 'vertical' | 'horizontal';

export type CustomControlsValue =
  string | number | (string | number)[] | undefined;

export interface CustomControlsFormData extends QueryFormData {
  controlType?: ControlType;
  filterColumn?: QueryFormColumn;
  orientation?: Orientation;
  includeAllOption?: boolean;
  multiSelect?: boolean;
  defaultValue?: string;
  hideTitle?: boolean;
  boldTitle?: boolean;
}

export interface CustomControlsTransformedProps {
  width: number;
  height: number;
  data: DataRecord[];
  controlType: ControlType;
  filterColumn?: QueryFormColumn;
  orientation: Orientation;
  includeAllOption: boolean;
  multiSelect: boolean;
  defaultValue: string;
  hideTitle: boolean;
  boldTitle: boolean;
  setDataMask: SetDataMaskHook;
  filterState?: FilterState;
}